Implement WOFF container
Only works if system zlib is used.

We simply synthesize a SFNT from the WOFF, create a Memory stream
for the new data, and load the SFNT as usual.

However, to get this to work I had to work around the fact that
most of the code redundantly passes stream around as function
parameters when it's already included in the face.  As such,
drivers that swap out stream need to be very careful.  Seems to
work now.

Does NOT add any API to access WOFF metadata or private blocks.