[threadsafety] Allocate TT_ExecContext in TT_Size instead of TT_Driver
Previously the code had stipulation for using a per-TT_Size exec
context if size->debug was true.  But there was no way that size->debug
could *ever* be true.  As such, the code was always using the singleton
TT_ExecContext that was stored in TT_Driver.  This was, clearly, not
threadsafe.

Remove size->debug.

Remove TT_Driver-global TT_ExecContext.

Lazy-create a TT_ExecContext for the TT_Size when needed.

With this patch loading glyphs from different faces from different
threads doesn't crash in the bytecode loader code.