[threadsafety] Do not reorder library->renderers upon use
Instead of keeping library->renderers in a MRU order, just leave
it as is.  The MRU machinery was an overengineering premature
optimization to begin with.  And wasn't thread-safe.

With this patch rasterizing glyphs from different faces from different
threads doesn't fail choosing rasterizer (FT_Err_Cannot_Render_Glyph).

Easiest to see that crash was to add a printf or otherwise let thread
yield in FT_Throw with debugging enabled.