Fix mono rasterizer uninitialized memory access
Apparently ras.cProfile might be uninitialized.  This will be the case
if ras.top == ras.cProfile->offset, as can be seen in End_Profile.
The overshoot code introduced in 72271140434028186a49a5dc5925f0727559e46f
violated this, accessing ras.cProfile->flags unconditionally just
before calling End_Profile (which then detected that cProfile is
uninitialized and didn't touch it).

This was harmless, and was not detected by valgrind before because the
objects were allocated on the raster_pool, which was always initialized.
With recent change to allocate raster buffers on the stack, valgrind now
reported this invalid access:

==29514== Conditional jump or move depends on uninitialised value(s)
==29514==    at 0x4E957CF: Render_Single_Pass (ftraster.c:1985)
==29514==    by 0x4E961CD: ft_black_render (ftraster.c:2955)
==29514==    by 0x4E9483F: ft_raster1_render (ftrend1.c:205)
==29514==    by 0x4E44B9D: FT_Render_Glyph_Internal (ftobjs.c:4098)
==29514==    by 0x4E99ED7: ftc_basic_family_load_bitmap (ftcbasic.c:141)
==29514==    by 0x4E99594: ftc_snode_load (ftcsbits.c:127)
==29514==    by 0x4E99818: ftc_snode_compare (ftcsbits.c:390)
==29514==    by 0x4E9B5AB: FTC_SBitCache_LookupScaler (ftcsbits.c:415)
==29514==    by 0x406BBB: FTDemo_Index_To_Bitmap (ftcommon.c:778)
==29514==    by 0x406D7F: FTDemo_Draw_Index (ftcommon.c:888)
==29514==    by 0x402F16: main (ftview.c:493)

Fix this by not accessing an uninitialized cProfile.